requirements.txt
# Core Framework
crewai>=0.80.0
crewai-tools>=0.12.0

# LLM Support
google-generativeai>=0.8.0

# HTTP & API
requests>=2.32.0
urllib3>=2.2.0

# Data Validation
pydantic>=2.10.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
pathlib>=1.0.1
